Ensure %CPU is available when disk quota disabled
The stat_collector task coded with the expectation that disk statistics
would always be available but when disk quotas are not enabled for
warden, the statistics are not populated and the collector would take an
exception dereferencing the nil message.  That exception would prevent
cpu statistics code from executing resulting in cpu data always
returning 0%.

This adds a check for nil on the disk usage assignment.